Do not use v1::Bound, use new ommx::Bound instead
